requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
google-generativeai==0.8.0
httpx==0.28.0
trafilatura==2.0.0
python-multipart==0.0.18
pillow>=11.2.1
beautifulsoup4==4.12.3
langdetect==1.0.9
PyMuPDF==1.25.0
python-dotenv==1.1.0
